==> 4_DIST_KERNEL.Rmd <==
---
title: "Script final"
author: "Borja Gómez"
date: "18/0362022"
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: true
    theme: paper
    fig_width: 6
    fig_height: 4
vignette: >
  %\VignetteIndexEntry{dataset}
  %\VignetteEngine{knitr::rmarkdown}
  \usepackage[utf8]{inputenc}
---

<style>
body {
text-align: justify}
</style>



```{r}
pacman::p_load(GOSemSim,GO.db, kdensity, ggplot2)
organismo = "org.Hs.eg.db"
GO_semsim = godata(organismo, ont="BP")
pvalor = 0.05
  ```

### SIN PROCESAR ###
```{r}
load(paste("analisis/enriquecimiento/terminos_significativos.rda", sep = "/"))


# Calculamos el IC para los términos detectados
valores_ic_fisher = na.omit((GO_semsim@IC)[names(go_term_sig[[1]])])
valores_ic_gsea = na.omit((GO_semsim@IC)[names(go_term_sig[[2]])])
valores_ic_elim = na.omit((GO_semsim@IC)[names(go_term_sig[[3]])])
valores_ic_weight = na.omit((GO_semsim@IC)[names(go_term_sig[[4]])])
valores_ic_pc = na.omit((GO_semsim@IC)[names(go_term_sig[[5]])])
valores_ic_weight01 = na.omit((GO_semsim@IC)[names(go_term_sig[[6]])])


datos = data.frame(valores = c(valores_ic_fisher,
                               valores_ic_gsea,
                               valores_ic_elim,
                               valores_ic_weight,
                               valores_ic_pc,
                               valores_ic_weight01),
                   nombre = c(rep("fisher",length(valores_ic_fisher)),
                              rep("gsea",length(valores_ic_gsea)),
                              rep("elim",length(valores_ic_elim)),
                              rep("weight",length(valores_ic_weight)),
                              rep("pc",length(valores_ic_pc)),
                              rep("weight01",length(valores_ic_weight01))))

datos <- datos[is.finite(datos$valores), ]

datos_normalizados <- (datos$valores - min(datos$valores)) / (max(datos$valores) - min(datos$valores))
datos$valores = datos_normalizados


valores_fisher <- subset(datos, nombre=="fisher")
valores_gsea <- subset(datos, nombre=="gsea")
valores_elim <- subset(datos, nombre=="elim")
valores_weight <- subset(datos, nombre=="weight")
valores_pc <- subset(datos, nombre=="pc")
valores_weight01 <- subset(datos, nombre=="weight01")
```


```{r}
svg(paste("analisis/graficos/KDE.svg", sep = "_"), width = 8, height = 6)
plot(kdensity(valores_fisher$valores),xlim = c(0,1.1), ylim = c(0,3.5), main = "",xlab="IC", lwd = 2, lty = 2 )
lines(kdensity(valores_gsea$valores), lty = 2, col = "blue", lwd = 2)
lines(kdensity(valores_elim$valores), lty = 1, col = "red", lwd = 2)
lines(kdensity(valores_weight$valores), lty = 1, col = "green", lwd = 2)
lines(kdensity(valores_pc$valores), lty = 1, col = "brown", lwd = 2)
lines(kdensity(valores_weight01$valores), lty = 1, col = "pink", lwd = 2)

legend("topleft",
legend = c("ORA", "GSEA", "elim", "weight", "PC", "weight01"),
col = c("black", "blue", "red", "green", "brown", "pink"),
lty = c(2,2,1,1,1,1),
cex=.75)

dev.off()
rm(list=setdiff(ls(), c("GO_semsim", "organismo")))
```

### PROCESADO ###
```{r}
load(paste("analisis/procesados/filtrado/terminos_significativos_filtrado.rda", sep = "/"))

# Calculamos el IC para los términos detectados
valores_ic_fisher = na.omit((GO_semsim@IC)[names(go_term_sig_wsc_filtrado[[1]])])
valores_ic_gsea = na.omit((GO_semsim@IC)[names(go_term_sig_wsc_filtrado[[2]])])
valores_ic_elim = na.omit((GO_semsim@IC)[names(go_term_sig_wsc_filtrado[[3]])])
valores_ic_weight = na.omit((GO_semsim@IC)[names(go_term_sig_wsc_filtrado[[4]])])
valores_ic_pc = na.omit((GO_semsim@IC)[names(go_term_sig_wsc_filtrado[[5]])])
valores_ic_weight01 = na.omit((GO_semsim@IC)[names(go_term_sig_wsc_filtrado[[6]])])

# Normalizamos los valores de IC para cada método
valores_ic_fisher <- (valores_ic_fisher - min(valores_ic_fisher)) / (max(valores_ic_fisher) - min(valores_ic_fisher))
valores_ic_gsea <- (valores_ic_gsea - min(valores_ic_gsea)) / (max(valores_ic_gsea) - min(valores_ic_gsea))
valores_ic_elim <- (valores_ic_elim - min(valores_ic_elim)) / (max(valores_ic_elim) - min(valores_ic_elim))
valores_ic_weight <- (valores_ic_weight - min(valores_ic_weight)) / (max(valores_ic_weight) - min(valores_ic_weight))
valores_ic_pc <- (valores_ic_pc - min(valores_ic_pc)) / (max(valores_ic_pc) - min(valores_ic_pc))
valores_ic_weight01 <- (valores_ic_weight01 - min(valores_ic_weight01)) / (max(valores_ic_weight01) - min(valores_ic_weight01))

# Generamos el plot de kernel combinando todos los métodos
svg(paste("analisis/graficos/KDE_filtrado.svg", sep = "/"), width = 8, height = 6)
plot(kdensity(valores_ic_fisher), ylim = c(0,3.5), main = "",xlab="IC", lwd = 2, lty = 2)
lines(kdensity(valores_ic_gsea), lty = 2, col = "blue", lwd = 2)
lines(kdensity(valores_ic_elim), lty = 1, col = "red", lwd = 2)
lines(kdensity(valores_ic_weight), lty = 1, col = "green", lwd = 2)
lines(kdensity(valores_ic_pc), lty = 1, col = "brown", lwd = 2)
lines(kdensity(valores_ic_weight01), lty = 1, col = "pink", lwd = 2)

legend("topleft",
       legend = c("ORA", "GSEA", "elim", "weight", "PC", "weight01"),
       col = c("black", "blue", "red", "green", "brown", "pink"),
       lty = c(2,2,1,1,1,1),
       cex=1)
dev.off()
#rm(list=setdiff(ls(), c("GO_semsim", "organismo")))
a<-lista_resultados[[2]]
```
